+++ skymap_scanner/server/reporter.py
@@ -6,11 +6,12 @@
 import bisect
 import dataclasses as dc
 import datetime as dt
-import functools
 import itertools
+import json
 import math
 import statistics
 import time
+from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional
 
 from rest_tools.client import RestClient
@@ -21,7 +22,7 @@ from ..utils import to_skyscan_result
 from ..utils.pixel_classes import NSidesDict
 from ..utils.utils import pyobj_to_string_repr
 from . import LOGGER
-from .utils import NSideProgression, httperror_silent_fail_or_infinite_retry
+from .utils import NSideProgression
 
 StrDict = Dict[str, Any]
 
@@ -222,6 +223,7 @@ class Reporter:
         n_posvar: int,
         nside_progression: NSideProgression,
         skydriver_rc: Optional[RestClient],
+        output_dir: Optional[Path],
         event_metadata: EventMetadata,
         predictive_scanning_threshold: float,
     ) -> None:
@@ -239,6 +241,8 @@ class Reporter:
                 - the list of nsides & pixel-extensions
             `skydriver_rc`
                 - a connection to the SkyDriver REST interface
+            `output_dir`
+                - a directory to write out results & progress
             `event_metadata`
                 - a collection of metadata about the event
             `predictive_scanning_threshold`
@@ -265,6 +269,7 @@ class Reporter:
         self._n_sent_by_nside: Dict[int, int] = {}
 
         self.skydriver_rc = skydriver_rc
+        self.output_dir = output_dir
         self.event_metadata = event_metadata
 
         # all set by calling initial_report()
@@ -370,11 +375,7 @@ class Reporter:
 
     def _get_result(self) -> SkyScanResult:
         """Get SkyScanResult."""
-        return to_skyscan_result.from_nsides_dict(
-            self.nsides_dict,
-            is_complete=self.is_event_scan_done,
-            event_metadata=self.event_metadata,
-        )
+        return to_skyscan_result.from_nsides_dict(self.nsides_dict, self.event_metadata)
 
     def predicted_total_recos(self) -> int:
         """Get a prediction for total number of recos for the entire scan.
@@ -588,23 +589,23 @@ class Reporter:
         }
 
         LOGGER.info(pyobj_to_string_repr(progress))
-        if not self.skydriver_rc:
-            return
 
         body = {
             "progress": progress,
             "event_metadata": dc.asdict(self.event_metadata),
             "scan_metadata": scan_metadata,
         }
-        await httperror_silent_fail_or_infinite_retry(
-            functools.partial(
-                self.skydriver_rc.request,
-                "PATCH",
-                f"/scan/{self.scan_id}/manifest",
-                body,
-            ),
-            self.is_event_scan_done,
-        )
+
+        if self.skydriver_rc:
+            await self.skydriver_rc.request(
+                "PATCH", f"/scan/{self.scan_id}/manifest", body
+            )
+
+        if self.output_dir:
+            fpath = self.output_dir / "scan-manifest.json"
+            with open(fpath, "w") as f:
+                json.dump(body, f, indent=4)
+            LOGGER.info(f"Manifest File (scan progress + metadata): {fpath}")
 
     async def _send_result(self) -> SkyScanResult:
         """Send result to SkyDriver (if the connection is established)."""
@@ -614,21 +615,14 @@ class Reporter:
         LOGGER.debug(  # don't log HUGE string
             f"Result info (# pixels per nside): {[(k, len(v)) for k, v in result.result.items()]}"
         )
-        if not self.skydriver_rc:
-            return result
 
-        body = {
-            "skyscan_result": serialized,
-            "is_final": self.is_event_scan_done,
-        }
-        await httperror_silent_fail_or_infinite_retry(
-            functools.partial(
-                self.skydriver_rc.request,
-                "PUT",
-                f"/scan/{self.scan_id}/result",
-                body,
-            ),
-            self.is_event_scan_done,
-        )
+        if self.skydriver_rc:
+            body = {"skyscan_result": serialized, "is_final": self.is_event_scan_done}
+            await self.skydriver_rc.request("PUT", f"/scan/{self.scan_id}/result", body)
+
+        if self.output_dir and result.result:  # don't write empty result to files
+            npz_fpath = result.to_npz(self.event_metadata, self.output_dir)
+            json_fpath = result.to_json(self.event_metadata, self.output_dir)
+            LOGGER.info(f"Result Files: {npz_fpath}, {json_fpath}")
 
         return result