+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 <!--next-version-placeholder-->
 
+## v3.12.0 (2023-08-17)
+
+### Other
+
+* <bot> update dependencies*.log files(s) ([`02d121c`](https://github.com/icecube/skymap_scanner/commit/02d121ce97257deb2e8cc2a84fed292a018a6b99))
+
+## v3.11.4 (2023-08-17)
+
+### Other
+
+* Use Script for CI Testing ([#209](https://github.com/icecube/skymap_scanner/issues/209)) ([`2ec587e`](https://github.com/icecube/skymap_scanner/commit/2ec587e90ef2d5bcd79015c63698192d9b38ee2f))
+
+## v3.11.3 (2023-07-21)
+
+### Other
+
+* Write Results & Manifest to File Throughout Scan ([#204](https://github.com/icecube/skymap_scanner/issues/204)) ([`200b1f5`](https://github.com/icecube/skymap_scanner/commit/200b1f599389a702b4e86364b0f7c8d707615f54))
+* CI: Bump `WIPACrepo/wipac-dev-py-setup-action@v2.2` ([`fac14d4`](https://github.com/icecube/skymap_scanner/commit/fac14d47970b5f6a5fee53736efe1db1488f2674))
+* <bot> update dependencies*.log files(s) ([`e2a9e15`](https://github.com/icecube/skymap_scanner/commit/e2a9e15351a2d6b80329e017f73e39525d398a6a))
+
 ## v3.11.2 (2023-07-21)
 
 ### Other